// src/enums/Misc.ts
export enum MiscEnum {
  GET_SECTIONS_BUTTON = 'Get Sections',
  GET_SECTIONS_ACTION_ID = 'get-sections',
  GET_PROJECTS_BUTTON = 'Get Projects',
  GET_PROJECTS_ACTION_ID = 'get-projects',
  PROJECT_ACTIONS_BLOCK = 'project-actions',
  CREATE_TASK_BUTTON = 'Create Task',
  CREATE_TASK_BUTTON_ACTION_ID = 'create-task-in-project',
  CREATE_TASK_IN_PROJECT_BUTTON = 'Create Task',
  CREATE_TASK_IN_PROJECT_BUTTON_ACTION_ID = 'create-task-in-project',
  VIEW_PROJECT_BUTTON = 'View',
  VIEW_PROJECT_ACTION_ID = 'view-project',
  SHARE_PROJECT_BUTTON = 'Share',
  SHARE_PROJECT_ACTION_ID = 'share-project',
  GET_TASKS_BUTTON = 'Get Tasks',
  GET_TASKS_ACTION_ID = 'get-tasks',
  TASK_ACTIONS_BLOCK = 'task-actions',
  VIEW_TASK_BUTTON = 'View',
  VIEW_TASK_ACTION_ID = 'view-task',
  EDIT_TASK_BUTTON = 'Edit',
  EDIT_TASK_ACTION_ID = 'edit-task',
  SHARE_TASK_BUTTON = 'Share',
  SHARE_TASK_ACTION_ID = 'share-task',
  DELETE_TASK_BUTTON = 'Delete',
  DELETE_TASK_ACTION_ID = 'delete-task',
  DELETE_SECTION_BUTTON = 'Delete',
  DELETE_SECTION_ACTION_ID = 'delete-section',
  DELETE_FOLDER_BUTTON = 'Delete',
  DELETE_FOLDER_ACTION_ID = 'delete-folder',
  DELETE_PROJECT_BUTTON = 'Delete',
  DELETE_PROJECT_ACTION_ID = 'delete-project',
  SECTION_ACTIONS_BLOCK = 'section-actions',
  SHARE_SECTION_BUTTON = 'Share Section',
  SHARE_SECTION_ACTION_ID = 'share-section',
  COMMENT_ACTIONS_BLOCK = 'comment-actions',
  SHARE_COMMENT_BUTTON = 'Share Comment',
  SHARE_COMMENT_ACTION_ID = 'share-comment',
  VIEW_COMMENT_BUTTON = 'View',
  VIEW_COMMENT_ACTION_ID = 'view-comment',
  DELETE_COMMENT_BUTTON = 'Delete',
  DELETE_COMMENT_ACTION_ID = 'delete-comment',
  GET_COMMENTS_BUTTON = 'Get Comments',
  GET_COMMENTS_ACTION_ID = 'get-comments',
  LABEL_ACTIONS_BLOCK = 'label-actions',
  SHARE_LABEL_BUTTON = 'Share Label',
  SHARE_LABEL_ACTION_ID = 'share-label',
  DELETE_LABEL_BUTTON = 'Delete Label',
  DELETE_LABEL_ACTION_ID = 'delete-label',
  CREATE_LABEL_BUTTON = 'Create Label',
  CREATE_LABEL_ACTION_ID = 'create-label',
  SHARED_LABEL_ACTIONS_BLOCK = 'shared-label-actions',
  RENAME_SHARED_LABEL_BUTTON = 'Rename',
  RENAME_SHARED_LABEL_ACTION_ID = 'rename-shared-label',
  REMOVE_SHARED_LABEL_BUTTON = 'Remove',
  REMOVE_SHARED_LABEL_ACTION_ID = 'remove-shared-label',
  CONVERT_TO_PERSONAL_LABEL_BUTTON = 'Convert to Personal',
  CONVERT_TO_PERSONAL_LABEL_ACTION_ID = 'convert-to-personal-label',
}